Fixed normal issue
Apparently the code is picky enough to not like specific filenames? I renamed the normal map from heightNormal2 to heightnormal, and it could be read out correctly. I want to assume it's because of capital letters or numbers, but it has no problem reading heightMap2, so I'm completely stumped.
Anyway it works now